Conflicts:
	install/win/installer.nsi
	install/win/installerlang.nsh
	src/Basescape/BasescapeState.cpp
	src/Basescape/SellState.cpp
	src/Battlescape/BattlescapeState.cpp
	src/Battlescape/DebriefingState.cpp
	src/Battlescape/NextTurnState.cpp
	src/Engine/Adlib/fmopl.cpp
	src/Engine/AdlibMusic.cpp
	src/Engine/AdlibMusic.h
	src/Engine/Game.cpp
	src/Geoscape/GeoscapeState.cpp
	src/Menu/AbandonGameState.cpp
	src/Menu/ListSaveState.cpp
	src/Menu/LoadGameState.cpp
	src/Menu/LoadGameState.h
	src/Menu/NewGameState.cpp
	src/Menu/NewGameState.h
	src/Menu/SaveGameState.cpp
	src/Menu/SaveGameState.h
	src/Ruleset/Ruleset.h
	src/Savegame/Base.cpp
	src/Savegame/SavedGame.cpp
	src/Savegame/Soldier.cpp
	src/Savegame/Soldier.h
	src/Savegame/Transfer.cpp
	src/Savegame/Transfer.h

+++ src/Basescape/BasescapeState.cpp
@@ -683,7 +683,7 @@ void BasescapeState::handleKeyPress(Action* action)
 
 		int key = action->getDetails()->key.keysym.sym;
 
-		for (int
+		for (size_t
 				i = 0;
 				i < MiniBaseView::MAX_BASES;
 				++i)
